requirements.txt
openai
feedparser
Pillow
requests
supabase
